+++ ntbs-service/Pages/Admin/Index.cshtml.cs
@@ -1,0 +1,12 @@
+﻿using Microsoft.AspNetCore.Mvc.RazorPages;
+
+namespace ntbs_service.Pages.Admin
+{
+    public class Index : PageModel
+    {
+        public void OnGet()
+        {
+            
+        }
+    }
+}
